+++ src/main/java/com/robertx22/age_of_exile/vanilla_mc/packets/DmgNumPacket.java
@@ -1,6 +1,6 @@
 package com.robertx22.age_of_exile.vanilla_mc.packets;
 
-import com.robertx22.age_of_exile.a_libraries.dmg_number_particle.OnDisplayDamage;
+import com.robertx22.age_of_exile.a_libraries.dmg_number_particle.DamageParticleAdder;
 import com.robertx22.age_of_exile.config.forge.ModConfig;
 import com.robertx22.age_of_exile.mmorpg.Ref;
 import com.robertx22.age_of_exile.uncommon.enumclasses.Elements;
@@ -17,12 +17,10 @@ public class DmgNumPacket extends MyPacket<DmgNumPacket> {
 
     public String element;
     public String string;
-    public double x;
-    public double y;
-    public double z;
-    public float height;
+
     public boolean isExp;
     public float number;
+    public int id;
 
     public DmgNumPacket() {
 
@@ -32,11 +30,8 @@ public class DmgNumPacket extends MyPacket<DmgNumPacket> {
 
         element = ele.toString();
         string = str;
-        x = entity.getX();
-        y = entity.getY();
-        z = entity.getZ();
-        height = entity.getHeight();
         this.number = number;
+        this.id = entity.getEntityId();
 
     }
 
@@ -48,25 +43,21 @@ public class DmgNumPacket extends MyPacket<DmgNumPacket> {
     @Override
     public void loadFromData(PacketByteBuf tag) {
         element = tag.readString();
-        x = tag.readDouble();
-        y = tag.readDouble();
-        z = tag.readDouble();
-        height = tag.readFloat();
         isExp = tag.readBoolean();
         string = tag.readString();
         number = tag.readFloat();
+        id = tag.readInt();
+
     }
 
     @Override
     public void saveToData(PacketByteBuf tag) {
         tag.writeString(element);
-        tag.writeDouble(x);
-        tag.writeDouble(y);
-        tag.writeDouble(z);
-        tag.writeFloat(height);
         tag.writeBoolean(isExp);
         tag.writeString(string);
         tag.writeFloat(number);
+        tag.writeInt(id);
+
     }
 
     @Override
@@ -76,7 +67,7 @@ public class DmgNumPacket extends MyPacket<DmgNumPacket> {
                 .sendMessage(new SText(Formatting.GREEN + "" + Formatting.BOLD + "+" + (int) number + " EXP"), false);
 
         } else if (isExp == false && ModConfig.get().client.dmgParticleConfig.ENABLE_FLOATING_DMG) {
-            OnDisplayDamage.displayParticle(element, string, x, y, z, height);
+            DamageParticleAdder.displayParticle(ctx.getPlayer().world.getEntityById(id), element, string);
         }
     }